Added fully functioning practice mode
-Added practice mode

Instead of just saving the player position, velocity and rotation, all objects in the scene (obstacles & boxes rotation, position, velocity, angular velocity and if rigidbody is active or not).

- Practice mode icons

Basically geometry dash, but with different keys (but I made the green diamond myself, so I can't be sued!)

- Rainbow trail

Not sure if it's working 100% tho

- Level 16

Now fully finished!

- Began working on Level 17

Level 1, but instead of the 'goal', the camera turns around and you go backwards (new mechanic)

- UI improvements

Animation fixes (such as enlarging when mouse is hovering over) and if no object is in focus and the keyboard is pressed, the last object is selected. (Still needs some work tho)

- Progress bar now is more flexible

Such as beeing able to turn around and decrease too

- Play the next level directly from the title screen

The next level is loaded 'additivly' to the title screen and when play is pressed, it switches to the level so you can play immidietly without having to go through the level select